Move file into lib directory
I'm not sure why I didn't do this in the first place so better to get it
done sooner rather than later.
