+++ lib/aqm_reading.rb
